archivist-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.3"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
sqlalchemy = "^2.0.49"
aiosqlite = "^0.22.1"
python-dateutil = "^2.9.0.post0"
python-multipart = "^0.0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
